refactor: remove unused components and update imports for better organization
