--- src/app/(toys)/hello-world/toy-layout/md/body.mdx
@@ -1,1 +1,0 @@
-作成したコンポーネント「ToysLayout」の使い方バックアップ。
