--- src/app/(toys)/hello-world/toy-layout/md/supplement.mdx
@@ -1,1 +1,0 @@
-補足テキスト
