+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
-import { AppHeader } from "@/components/molecules/app-header";
-import { AppSidebar } from "@/components/molecules/app-sidebar";
+import { AppHeader } from "@/components/layout/app-header";
+import { AppSidebar } from "@/components/layout/app-sidebar";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import type { Metadata } from "next";
 import { Noto_Sans_JP } from "next/font/google";
